feat: 4 ways to parallel run
